+++ AvespoirTest.Core/Modules/Commands/PublicCommands/Template.cs
@@ -1,12 +1,8 @@
 ﻿using AvespoirTest.Core.Attributes;
-using System.Threading.Tasks;
-using System;
-using DSharpPlus.Entities;
-using DSharpPlus.Exceptions;
-using AvespoirTest.Core.Configs;
 using AvespoirTest.Core.Modules.Logger;
-using AvespoirTest.Core.Exceptions;
 using AvespoirTest.Core.Modules.Utils;
+using System;
+using System.Threading.Tasks;
 
 namespace AvespoirTest.Core.Modules.Commands {