+++ AvespoirTest.Core/Modules/Events/ReadyEvent.cs
@@ -1,11 +1,7 @@
-﻿using AvespoirTest.Core.Database;
-using AvespoirTest.Core.Modules.Logger;
+﻿using AvespoirTest.Core.Modules.Logger;
 using DSharpPlus;
 using DSharpPlus.Entities;
 using DSharpPlus.EventArgs;
-using MongoDB.Driver;
-using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace AvespoirTest.Core.Modules.Events {
@@ -20,18 +16,20 @@ namespace AvespoirTest.Core.Modules.Events {
 			Name = "Bot Ready!",
 		};
 
-		internal static async Task Main(ReadyEventArgs ReadyEventObjects) {
+		internal static Task Main(ReadyEventArgs ReadyEventObjects) {
 			new DebugLog("ReadyEvent " + "Start...");
 			BaseDiscordClient ClientBase = ReadyEventObjects.Client;
-			await Client.Bot.UpdateStatusAsync(StartingStatus, UserStatus.DoNotDisturb);
+			Client.Bot.UpdateStatusAsync(StartingStatus, UserStatus.DoNotDisturb).ConfigureAwait(false);
 
-			await Client.Bot.UpdateStatusAsync(ReadyStatus, UserStatus.Online);
+			Client.Bot.UpdateStatusAsync(ReadyStatus, UserStatus.Online).ConfigureAwait(false);
 
 			new InfoLog($"{ClientBase.CurrentUser.Username} Bot Ready!");
-			await Task.Delay(5000).ConfigureAwait(false);
+			Task.Delay(5000).ConfigureAwait(false);
 
-			await Client.Bot.UpdateStatusAsync(null, UserStatus.Online);
+			Client.Bot.UpdateStatusAsync(null, UserStatus.Online).ConfigureAwait(false);
 			new DebugLog("ReadyEvent " + "End...");
+
+			return Task.CompletedTask;
 		}
 	}
 }